sched/fair: Add more core cookie check in wake up fast path
In wake up fast path, the changes will check for core cookie
match before settling for a cpu.

This is to avoid a task waking up on a cpu with incompatible cookie
as much as possible. The reasoning is even if the cpu is idle, it is
better to avoid it when the sibling cpu is running an task cannot be
scheduled together, forcing one into idle.

Signed-off-by: Xiaoyi Chen <cxiaoyi@amazon.com>
Reviewed-by: Alan Powell <alpowell@amazon.com>
Reviewed-by: Eitan Eliahu <eeliahu@amazon.com>
Reviewed-by: Full Name <goossen@amazon.tld>
Reviewed-by: Roland Paterson-Jones <rolandp@amazon.com>