sched/core: Stealing more frequently
When the idle task runs both due to normal idle or forced idle, the core
balancer is triggered. In the core balancer we early out if the HT is
not forced-idle.

This is interesting because in the truly idle case we don't do any
balancing but as soon as there is one other task running we aggressively
try to steal other tasks (each time the HT switches between idle and
non-idle task). So running this somewhat frequently on an idle HT is
not a big deal other than that it may mess with power savings.

Following the logic in the periodic balancer using a jiffies timeout we
can run the core balancer once every N jiffies or if it is truly forced
idle.

The below code runs the balancer every 1s.

Signed-off-by: Xiaoyi Chen <cxiaoyi@amazon.com>
Reviewed-by: Alan Powell <alpowell@amazon.com>
Reviewed-by: Full Name <goossen@amazon.tld>
Reviewed-by: Roland Paterson-Jones <rolandp@amazon.com>