sched: avoid alb for tasks in core affine group
Don't kick the active_load_balance_cpu_stop, if the curr task on
busiest CPU belongs to a core affine group.

We currently don't have a metrics to decide whether migrating all
tasks in a core affine group with alb. Avoid alb for these tasks as
a temporary solution to keep tasks together.

Signed-off-by: Xiaoyi Chen <cxiaoyi@amazon.com>
Reviewed-by: Alan Powell <alpowell@amazon.com>
Reviewed-by: Roland Paterson-Jones <rolandp@amazon.com>