sched: Improve secondary core task picking
Consider all tasks with a matching cookie, when making the secondary
decision during core picking. Allow to discard even the best task, when
it is not suitable for execution.

At the very least, this is important for CFS tasks, as the core tree,
which serves as a basis for the seconday task selection, does not reflect
the dynamicy of vruntime or other CFS runtime decisions. Instead, the core
scheduler happily picks the same task over and over, which may cause it
to overexecute which potentially results in starvation effects because
the task consumes all the available bandwidth of its group, or because
its vruntime grows unbounded setting it up for later starving when other
tasks are eventually allowed to catch up.

Even with considering all matching tasks, the best task may still not be
suitable, either because its vruntime is already too far ahead or it is
currently throttled. Hence, the need to even discard the best task.

Notes:

We do not yet perfectly match the original CFS task selection,
as cfs_prio_less() does not account for buddies for example.

There might still be an issue with selecting a task that is about
to be throttled.

Remaining issues (brain dump):

Load looks much more reasonable now. It is somewhat deterministic, but
not yet fully stable for certain situations, because the primary core
task picking is too arbitrary. Sometimes it also picks an unfair stable
situation.

Example 1:
- cookie A: two tasks on one sibling, one on the other
- cookie B: one task per sibling

   SMT1 SMT2   SMT1 SMT2   SMT1 SMT2
A  33%  33%    25%  50%    30%  40%
   33%  --     25%         30%  --
B  33%  33%    25%  50%    30%  40%
               --          --

All three load sitations may happen, they all are currently valid
and considered fair. Left one favors SMT1 in primary selection
(and is the one I consider fair), middle one favors SMT2. Right
one is a mixture.

Example 2:
- cookie A: one nice-10 task on SMT1, one normal task on SMT2
- cookie B: same
- cookie C: one nice-10 task on SMT1

   SMT1 SMT2   SMT1 SMT2   SMT1 SMT2
A  33%  33%    20%  40%    5%   47%
               --          --
B  33%  33%    20%  40%    5%   47%
               --          --
C  33%         20%  --     5%   --

Left favors SMT1, right favors SMT2 (and happens to be the fair variant).

Example 3:
- cookie A: one task on SMT2
- cookie B: one nice-10 task per sibling

   SMT1 SMT2   SMT1 SMT2   SMT1 SMT2
A  --   75%    --   66%    --   90%
B  25%   7%    33%   7%    10%  10%
        --          --

Left/middle is what happens (favoring SMT1/favoring SMT2), because during
execution of B, the core scheduler forgets the vruntime offset between
the runqueues. Right is what should happen.

Note, that the force-idle mechanism seems to cause fairness issues in
itself, when the prio_less() in the primary selection is suddenly
confronted with a task from a different task group. (somewhat unverified)

Also (unverified), from code analysis, the transition to un-cookied tasks
looks as if it may force-idle a sibling or not notify a sibling or cause
the sibling to perform a core wide selection right after.

Stuff for later:

Load balancing with cookies.
Deadline scheduling with cookies.
Round-robin RT tasks with cookies. (not reflected in prio_less())
RT tasks with cookies in general (groups, balancing)

Upstream-status: Upstreamable after further work to complete functionality
Patch-revision: 1

Signed-off-by: Jan H. Schönherr <jschoenh@amazon.de>
Reviewed-by: Alan Powell <alpowell@amazon.com>
Reviewed-by: Roland Paterson-Jones <rolandp@amazon.com>

CR: https://code.amazon.com/reviews/CR-74911929