sched/fair: Core sched always picks long waiters
There are scenarios where CFS heuristics fail to meet the goal of
allocating a slice of task for each scheduling period. This can
specifically happen with core picking which throws a big spanner in CFS wheels
because it picks the best tasks across siblings and then secondary cores
just have to follow the lead. Additionally a core can become force idle, which
just freezes the time for that cpu (vruntime stops); there exist some
mitigations to consider force idle but they are reset as soon as the force
idle stops, even for a short amount of time. Generally tasks joining and
leaving the competition make vruntime placement heuristics critical and the
result of all of the above is that there's no strong guarantees on the run
delay that a running task might experience.

To blanket address these issues for fair tasks, we introduce the concept of
absolute run delay in task prioritization. If a task experiences a run delay
that is high enough (dynamic sched latency target, the period during which each
fair task should get a slice, i.e 24ms or higher under elevated contention), we
will override vruntime considerations and pick that task instead.

Signed-off-by: Fernand Sieber <sieberf@amazon.com>
Reviewed-by: Demetre Liakos <dlliakos@amazon.com>
Reviewed-by: Mpho Mathabatha <mphoma@amazon.com>
Reviewed-by: Himal Rama <ramhimal@amazon.com>
Reviewed-by: Roland Paterson-Jones <rolandp@amazon.com>