sched/fair: Allow tasks to run on secondary core when not throttled
On "sched: Improve secondary core task picking" (43a291d3ba5c7) we have
introduced some logic to a) compare secondary core candidate picking and
pick the best one and b) put secondary core in forced idle if the best task
is too far ahead in vruntime.

With the introduction of the long waiters picking in the previous commit,
it is not necessary anymore to put the secondary core in forced idle as the
task that was temporarily put ahead in vruntime will still be picked
occasionally as it becomes a long waiter and keep make some progress until
other tasks catch up.

Signed-off-by: Fernand Sieber <sieberf@amazon.com>
Reviewed-by: Demetre Liakos <dlliakos@amazon.com>
Reviewed-by: Mpho Mathabatha <mphoma@amazon.com>
Reviewed-by: Himal Rama <ramhimal@amazon.com>
Reviewed-by: Roland Paterson-Jones <rolandp@amazon.com>