sched: Remove dependency of long waiter picks on sched period
he replacement of WF2Q by EEVDF means that the scheduling algorithm is no
longer modeling a scheduler period divided in fair slices. As such, the
concept of a scheduler period doesn't exist anymore. Long waiters picking
logic introduced in previous downstream commits use of the scheduler period
thus needs to be updated to enable migration to EEVDF.

There are two use cases for scheduler period in long waiters.

The first is to provide a minimal threshold at which a task run delay may
be considered to be excessive. Instead of setting this threshold to the
scheduler period (in WF2Q any task is supposed to get a fair slice within
the period), we use the heuristic that any runnable task should get the
mimimum slice. Of course this doesn't account for fairness as it would be
expected that a light task sees larger run delays, but this consdieration
is already implemented as separate checks.

The second use case is to override the task comparison logic between two
picks from different CPUs: we should prefer the task with the largest run
delay if the task is a long waiter. This was previously implemented using
the heuristic that the task run delay was larger than the sched period.
Instead of using a heuristic, we introduce the concept of pick type which
allows us to remember if a a task was picked because if it was a long
waiter or not. This information can then be reused when comparing picks
from different CPUs.

Signed-off-by: Fernand Sieber <sieberf@amazon.com>